Added a writable wrapper for live query
It might not be performant or robust, but within the supported scope it does what I wanted it to do.

Modifying the store value will cause it to put the new content into the Database in the specified table. It also allows the user to provide a "default" value which will be presented to subscribers if the live query fails to provide a value (or has yet to return the first value).

In this way, it allows for other liveQueries watching the same content to update alongside writes to this writable.

All in all, it means I no longer need to share a single writable reference across pages or components. I can instantiate multiple liveQueries in isolation, and they'll all update with the changes. Thus eliminating a lot of cross dependency in the code.